--- a/hw10_program_optimization/stats_optimization_test.go
+++ b/hw10_program_optimization/stats_optimization_test.go
@@ -48,6 +48,24 @@ func TestGetDomainStat_Time_And_Memory(t *testing.T) {
 	require.Less(t, mem, memoryLimit, "the program is too greedy")
 }
 
+func BenchmarkGetDomainStat(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		r, err := zip.OpenReader("testdata/users.dat.zip")
+		require.NoError(b, err)
+		defer r.Close()
+
+		require.Equal(b, 1, len(r.File))
+		data, err := r.File[0].Open()
+		require.NoError(b, err)
+
+		stat, err := GetDomainStat(data, "biz")
+		require.NoError(b, err)
+		require.Equal(b, expectedBizStat, stat)
+
+		r.Close()
+	}
+}
+
 var expectedBizStat = DomainStat{
 	"abata.biz":         25,
 	"abatz.biz":         25,