@bardex 2021-09-30
Before Benchmarks:
```
goos: linux
goarch: amd64
pkg: github.com/bardex/otus_homework/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-6700T CPU @ 2.80GHz
BenchmarkGetDomainStat-8   	       1	1224237703 ns/op	305172568 B/op	 3046718 allocs/op
PASS
ok  	github.com/fixme_my_friend/hw10_program_optimization	1.235s
```
After Benchmarks:
```
goos: linux
goarch: amd64
pkg: github.com/bardex/otus_homework/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-6700T CPU @ 2.80GHz
BenchmarkGetDomainStat-8   	       4	 293931132 ns/op	11357332 B/op	 1424397 allocs/op
PASS
ok  	github.com/bardex/otus_homework/hw10_program_optimization	2.359s
```
Benchstat:
```
name             before time/op    after time/op    delta
GetDomainStat-8     1.22s ± 0%     0.29s ± 0%  -75.99%

name             before alloc/op   after alloc/op   delta
GetDomainStat-8     305MB ± 0%      11MB ± 0%  -96.28%

name             before allocs/op  after allocs/op  delta
GetDomainStat-8     3.05M ± 0%     1.42M ± 0%  -53.25%
```