@farir1408 2021-10-12
Здравствуйте!

Спасибо за выполнение домашнего задания.
Отличная работа, принято 10/10.